move to the core dir, add root
